requirements.txt
Flask==3.0.2
Flask-SocketIO==5.3.6
Flask-Cors==4.0.0
cmake==3.28.3
amazon-transcribe==0.6.2
SpeechRecognition==3.10.1
boto3==1.34.62
pillow==10.2.0